+++ capacitor.config.ts
@@ -3,7 +3,7 @@ import type { CapacitorConfig } from '@capacitor/cli';
 const config: CapacitorConfig = {
   appId: 'com.tenkey.domacinko',
   appName: 'Domacinko',
-  webDir: '.',
+  webDir: 'www',
   server: {
     androidScheme: 'https',
     cleartext: false